+++ R/combine_widgets.R
@@ -129,6 +129,7 @@
 #' }
 #'
 #' @export
+#' @importFrom htmltools tagGetAttribute
 combineWidgets <- function(..., list = NULL, nrow = NULL, ncol = NULL, title = NULL,
                            rowsize = 1, colsize = 1, byrow = TRUE,
                            titleCSS = "",
@@ -152,12 +153,25 @@ combineWidgets <- function(..., list = NULL, nrow = NULL, ncol = NULL, title = N
     preRenderHook = preRenderCombinedWidgets
   )
 
-  # Add dependencies of embedded widgets
-  deps <- lapply(widgets, function(x) {
-    if (is.null(attr(x, "package"))) return(NULL)
-    append(getDependency(class(x)[1], attr(x, "package")), x$dependencies)
-  })
-  deps <- do.call(c, deps)
+  # Add dependencies of embedded widgets or shiny tags
+  # This works through the widgets recursively, in case
+  # we were passed a shiny.tag.list or other list of
+  # non-widgets.
+
+  getDeps <- function(x) {
+    if (!is.null(attr(x, "package")))
+      append(tryCatch(getDependency(class(x)[1], attr(x, "package")),
+                      error = function(e) NULL), x$dependencies)
+    else if (!is.null(attr(x, "html_dependencies")))
+      attr(x, "html_dependencies")
+    else if (is.list(x))
+      do.call(c, lapply(x, getDeps))
+  }
+  deps <- c(getDeps(widgets),
+            getDeps(header),
+            getDeps(footer),
+            getDeps(leftCol),
+            getDeps(rightCol))
 
   res$dependencies <- deps
 
@@ -212,6 +226,9 @@ renderCombineWidgets <- function(expr, env = parent.frame(), quoted = FALSE) {
 
 # Private function used to prerender a combinedWidgets object
 preRenderCombinedWidgets <- function(x) {
+
+  hasCrosstalkInputs <- any(unlist(lapply(x$widgets, isCrosstalkInput)))
+
   widgets <- lapply(unname(x$widgets), function(w) {
     if (is.atomic(w)) return(structure(list(x = as.character(w)), class = "html"))
     if (is.null(w$preRenderHook)) {
@@ -249,19 +266,29 @@ preRenderCombinedWidgets <- function(x) {
     })
   }
 
+  # Get the HTML class for each widget, plus "cw-widget"
+  elementClass <- sapply(widgets[1:ncells], function(w) {
+    result <- NULL
+    if (inherits(w, "htmlwidget"))
+      result <- class(w)[1]
+    else if (inherits(w, "shiny.tag"))
+      result <- tagGetAttribute(w, "class")
+    paste(result, "cw-widget")
+  })
 
   # Construct the html of the combined widget
   dirClass <- ifelse(x$params$byrow, "cw-by-row", "cw-by-col")
 
   widgetEL <- mapply(
-    function(id, size) {
+    function(id, size, class) {
       sprintf('<div class="cw-col" style="flex:%s;-webkit-flex:%s">
-              <div id="%s" class="cw-widget" style="width:100%%;height:100%%"></div>
+              <div id="%s" class="%s" style="width:100%%;height:100%%"></div>
               </div>',
-              size, size, id)
+              size, size, id, class)
     },
     id = elementId,
-    size = rep(colsize, length.out = ncells)
+    size = rep(colsize, length.out = ncells),
+    class = elementClass
   )
 
   rowsEl <- lapply(1:nrow, function(i) {
@@ -295,7 +322,17 @@ preRenderCombinedWidgets <- function(x) {
   data <- lapply(widgets, function(w) w$x)
   widgetType <- sapply(widgets, function(w) class(w)[1])
 
-  x$x <- list(data = data, widgetType = widgetType, elementId = elementId, html = html);
+  x$x <- list(data = data, widgetType = widgetType, elementId = elementId, html = html,
+              hasCrosstalkInputs = hasCrosstalkInputs);
 
   x
+}
+
+# Check whether a widget is a crosstalk-package input, which will need special
+# initialization within combineWidgets()
+
+isCrosstalkInput <- function(w) {
+  inherits(w, "shiny.tag") &&
+    !is.null(w$attribs) &&
+    grepl("crosstalk-input", w$attribs$class)
 }