+++ R/inputs.R
@@ -287,6 +287,65 @@ mwSelect <- function(choices = value, value = NULL, label = NULL, ...,
   )
 }
 
+#' Add a Select list input to a manipulateWidget gadget
+#'
+#' @param choices
+#'   Vector or list of choices. If it is named, then the names rather than the
+#'   values are displayed to the user.
+#' @param value
+#'   Initial value of the input. If not specified, the first choice is used.
+#' @param ...
+#'   Other arguments passed to function\code{\link[shiny]{selectInput}}.
+#' @param multiple
+#'   Is selection of multiple items allowed?
+#' @param options
+#'   A list of options. See the documentation of selectize.js for possible options
+#' @inheritParams mwSlider
+#'
+#' @return
+#' A function that will generate the input control.
+#'
+#' @examples
+#' if (require(plotly)) {
+#'   mydata <- data.frame(x = 1:100, y = rnorm(100))
+#'
+#'   # Select multiple values
+#'   manipulateWidget(
+#'     {
+#'       if (length(species) == 0) mydata <- iris
+#'       else mydata <- iris[iris$Species %in% species,]
+#'
+#'       plot_ly(mydata, x = ~Sepal.Length, y = ~Sepal.Width,
+#'               color = ~droplevels(Species), type = "scatter", mode = "markers")
+#'     },
+#'     species = mwSelectize(c("Select one or two species : " = "", levels(iris$Species)),
+#'         multiple = TRUE, options = list(maxItems = 2))
+#'   )
+#' }
+#'
+#' @export
+#' @family controls
+mwSelectize <- function(choices = value, value = NULL, label = NULL, ...,
+                     multiple = FALSE, options = NULL, .display = TRUE) {
+  params <- dotsToExpr()
+  params$choices <- substitute(choices)
+  params$multiple <- substitute(multiple)
+  params$options <- substitute(options)
+  value <- substitute(value)
+  Input(
+    type = "select", value = value, label = label, params = params,
+    display = substitute(.display),
+    validFunc = function(x, params) {
+      x <- intersect(x, unlist(params$choices))
+      if (params$multiple) return(x)
+      else if (length(x) > 0) return(x[1])
+      else return(params$choices[[1]])
+    },
+    htmlFunc = htmlFuncFactory(shiny::selectizeInput, "selected"),
+    htmlUpdateFunc = changeValueParam(shiny::updateSelectizeInput, "selected")
+  )
+}
+
 #' Add a checkbox to a manipulateWidget gadget
 #'
 #' @param value
@@ -601,7 +660,7 @@ mwSharedValue <- function(expr = NULL) {
 #' @param ... inputs that will be grouped in the box
 #' @param .display expression that evaluates to TRUE or FALSE, indicating when
 #'   the group should be shown/hidden.
-#'
+#' @param label label of the group inputs
 #' @return Input of type "group".
 #'
 #' @examples
@@ -621,14 +680,14 @@ mwSharedValue <- function(expr = NULL) {
 #'
 #' @export
 #' @family controls
-mwGroup <- function(..., .display = TRUE) {
+mwGroup <- function(..., label = NULL, .display = TRUE) {
   inputs <- list(...)
   if (is.null(names(inputs))) stop("All arguments need to be named.")
   for (i in inputs) if (!inherits(i, "Input")) stop("All arguments need to be Input objects.")
 
   Input(
     type = "group", value = list(...), params = list(),
-    display = substitute(.display),
+    label = label, display = substitute(.display),
     htmlFunc = function(id, label, value, params, ns) {
       htmlElements <- lapply(value, function(x) x$getHTML(ns))